Reorg search files utilities
