+++ gpm/tests/test_bucket/test_io.py
@@ -27,15 +27,10 @@
 """This module tests the bucket I/O utilities."""
 import os
 
-import pytest
-
 from gpm.bucket.io import (
-    get_filepaths_by_path,
-    get_filepaths_within_paths,
+    get_filepaths,
+    get_filepaths_by_partition,
     get_partitions_paths,
-    get_subdirectories,
-    search_leaf_directories,
-    search_leaf_files,
     write_bucket_info,
 )
 from gpm.bucket.partitioning import LonLatPartitioning
@@ -69,114 +64,6 @@ def create_test_bucket(bucket_dir):
             f.write("")  # Writing an empty string to create the file
 
 
-def test_get_filepaths_within_paths(tmp_path):
-    """Test get_filepaths_within_paths."""
-    # Create the directory structure
-    bucket_dir = tmp_path
-    create_test_bucket(bucket_dir=bucket_dir)
-    paths = [
-        os.path.join(bucket_dir, "lon_bin=-5.0", "lat_bin=-5.0"),
-        os.path.join(bucket_dir, "lon_bin=-5.0", "lat_bin=5.0"),
-    ]
-
-    # Test results without filtering
-    filepaths_p = get_filepaths_within_paths(paths, parallel=True)
-    filepaths = get_filepaths_within_paths(paths, parallel=False)
-    assert len(filepaths) == 5
-    assert filepaths == filepaths_p
-
-    # Test results with extension filtering
-    filepaths_p = get_filepaths_within_paths(paths, parallel=True, file_extension=".parquet")
-    filepaths = get_filepaths_within_paths(paths, parallel=False, file_extension=".parquet")
-    assert len(filepaths) == 4
-    assert filepaths == filepaths_p
-
-    # Test results with glob filtering
-    filepaths_p = get_filepaths_within_paths(paths, parallel=True, glob_pattern="*.V07B_*")
-    filepaths = get_filepaths_within_paths(paths, parallel=False, glob_pattern="*.V07B_*")
-    assert len(filepaths) == 1
-    assert filepaths == filepaths_p
-
-    # Test results with regexp
-    filepaths_p = get_filepaths_within_paths(paths, parallel=True, regex_pattern="2B\\.GPM.*\\.parquet$")
-    filepaths = get_filepaths_within_paths(paths, parallel=False, regex_pattern="2B\\.GPM.*\\.parquet$")
-    assert len(filepaths) == 1
-    assert filepaths == filepaths_p
-
-
-def test_get_filepaths_by_path(tmp_path):
-    """Test get_filepaths_by_path."""
-    # Create the directory structure
-    bucket_dir = tmp_path
-    create_test_bucket(bucket_dir=bucket_dir)
-    path1 = os.path.join(bucket_dir, "lon_bin=-5.0", "lat_bin=-5.0")
-    path2 = os.path.join(bucket_dir, "lon_bin=-5.0", "lat_bin=5.0")
-    paths = [path1, path2]
-
-    # Test results without filtering
-    dict_filepaths_p = get_filepaths_by_path(paths, parallel=True)
-    dict_filepaths = get_filepaths_by_path(paths, parallel=False)
-    assert len(dict_filepaths) == 2
-    assert dict_filepaths == dict_filepaths_p
-    assert len(dict_filepaths[path1]) == 2
-    assert len(dict_filepaths[path2]) == 3
-
-    # Test results with filtering
-    dict_filepaths_p = get_filepaths_by_path(paths, parallel=True, file_extension=".parquet", glob_pattern="*.V07B_*")
-    dict_filepaths = get_filepaths_by_path(paths, parallel=False, file_extension=".parquet", glob_pattern="*.V07B_*")
-    assert len(dict_filepaths) == 2
-    assert dict_filepaths == dict_filepaths_p
-    assert len(dict_filepaths[path1]) == 0
-    assert len(dict_filepaths[path2]) == 1
-    assert dict_filepaths[path1] == []
-
-    # Test results filtering with regexp
-    dict_filepaths_p = get_filepaths_by_path(paths, parallel=True, regex_pattern=r"2B\.GPM.*\.parquet$")
-    dict_filepaths = get_filepaths_by_path(paths, parallel=False, regex_pattern=r"2B\.GPM.*\.parquet$")
-    assert len(dict_filepaths) == 2
-    assert dict_filepaths == dict_filepaths_p
-    assert len(dict_filepaths[path1]) == 0
-    assert len(dict_filepaths[path2]) == 1
-    assert dict_filepaths[path1] == []
-
-
-@pytest.mark.parametrize("remove_base_path", [True, False])
-def test_search_leaf_directories(tmp_path, remove_base_path):
-    # Create the directory structure
-    bucket_dir = tmp_path
-    create_test_bucket(bucket_dir=bucket_dir)
-    leaf_path1 = os.path.join("lon_bin=-5.0", "lat_bin=-5.0")
-    leaf_path2 = os.path.join("lon_bin=-5.0", "lat_bin=5.0")
-    path1 = os.path.join(bucket_dir, leaf_path1)
-    path2 = os.path.join(bucket_dir, leaf_path2)
-
-    paths = [path1, path2]
-    leaf_paths = [leaf_path1, leaf_path2]
-
-    # Test results without filtering
-    leaf_directories_p = search_leaf_directories(base_dir=bucket_dir, parallel=True, remove_base_path=remove_base_path)
-    leaf_directories = search_leaf_directories(base_dir=bucket_dir, parallel=False, remove_base_path=remove_base_path)
-    assert len(leaf_directories) == 2
-    assert leaf_directories == leaf_directories_p
-    if remove_base_path:
-        assert sorted(leaf_directories) == sorted(leaf_paths)
-    else:
-        assert sorted(leaf_directories) == sorted(paths)
-
-
-def test_get_subdirectories(tmp_path):
-    # Create the directory structure
-    bucket_dir = tmp_path
-    create_test_bucket(bucket_dir=bucket_dir)
-
-    # Test results
-    results = get_subdirectories(base_dir=bucket_dir, path=False)
-    assert results == ["lon_bin=-5.0"]
-    results = get_subdirectories(base_dir=os.path.join(bucket_dir, "lon_bin=-5.0"), path=False)
-    expected_results = ["lat_bin=5.0", "lat_bin=-5.0"]
-    assert sorted(results) == sorted(expected_results)
-
-
 def test_get_partitions_paths(tmp_path):
     # Create the directory structure
     bucket_dir = tmp_path
@@ -190,19 +77,44 @@ def test_get_partitions_paths(tmp_path):
     assert sorted(results) == sorted(expected_results)
 
 
-def test_search_leaf_files_in_parallel(tmp_path):
+def test_get_filepaths(tmp_path):
     # Create the directory structure
     bucket_dir = tmp_path
     create_test_bucket(bucket_dir=bucket_dir)
-
     # Test results without filtering
-    filepaths_p = search_leaf_files(
-        base_dir=bucket_dir,
+    results = get_filepaths(
+        bucket_dir=bucket_dir,
         parallel=True,
+        file_extension=None,
+        glob_pattern=None,
+        regex_pattern=None,
     )
-    filepaths = search_leaf_files(
-        base_dir=bucket_dir,
-        parallel=False,
+    assert len(results) == 5
+    # Test results with filtering
+    results = get_filepaths(
+        bucket_dir=bucket_dir,
+        parallel=True,
+        file_extension="parquet",
+        glob_pattern=None,
+        regex_pattern=None,
     )
-    assert len(filepaths) == 5
-    assert sorted(filepaths) == sorted(filepaths_p)
+    assert len(results) == 4
+
+
+def test_get_filepaths_by_partition(tmp_path):
+    # Create the directory structure
+    bucket_dir = tmp_path
+    create_test_bucket(bucket_dir=bucket_dir)
+    # Test results without filtering
+    dict_results = get_filepaths_by_partition(
+        bucket_dir=bucket_dir,
+        parallel=True,
+        file_extension=None,
+        glob_pattern=None,
+        regex_pattern=None,
+    )
+    assert isinstance(dict_results, dict)
+    expected_keys = [f"lon_bin=-5.0{os.sep}lat_bin=-5.0", f"lon_bin=-5.0{os.sep}lat_bin=5.0"]
+    assert sorted(dict_results) == sorted(expected_keys)
+    assert len(dict_results[expected_keys[0]]) == 2
+    assert len(dict_results[expected_keys[1]]) == 3
